RenderArea: use [superview bounds] and not frame
we must resize the render area to cover the full superview,
which is the contentsView of the NSWindow, but not have
the same origin!

It happens to be identical on Cocoa, but not on current
mySTEP which uses an NSThemwFrame with flipped coordinates.

Therefore the original code did work on Cocoa but moved the
render area up by the title bar width.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>